Header menu

+++ pages/api/article/[id].ts
@@ -3,6 +3,8 @@ import Article from "@/models/ArticleModel";
 import { NextApiRequest, NextApiResponse } from "next";
 import InventoryLocation from "@/models/InventoryLocationModel";
 import Vehicle from "@/models/VehicleModel";
+import { todayDate } from "@/lib/setDate";
+import TransactionHistory from "@/models/TransactionHistoryModel";
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
@@ -42,15 +44,45 @@ export default async function handler(
 
     case "DELETE":
       try {
-        const deletedLocation = await Article.deleteOne({ _id: id });
-        if (deletedLocation.deletedCount < 1) {
+        const findArticle = await Article.findById(id);
+
+        if (!findArticle) {
           return res
             .status(400)
-            .json({ success: false, data: "Article not deleted" });
+            .json({ success: false, data: "Artikel ej borttagen" });
         }
+
+        const createTransactionHistory = {
+          direction: "-",
+          cause: "Artikel permanent borttagen",
+          qty: findArticle.qty,
+          article: findArticle,
+          comment: "",
+          createdDate: todayDate,
+        };
+
+        const transactionHistory = await TransactionHistory.create(
+          createTransactionHistory
+        );
+
+        if (!transactionHistory) {
+          return res.status(400).json({
+            success: false,
+            data: "Något gick fel, artikel ej borttagen",
+          });
+        }
+
+        const deletedLocation = await Article.deleteOne({ _id: id });
+        if (deletedLocation.deletedCount < 1) {
+          return res.status(400).json({
+            success: false,
+            data: "Något gick fel, artikel ej borttagen",
+          });
+        }
+
         res.status(200).json({
           success: true,
-          data: "Article is successfully deleted",
+          data: "Artikeln är borttagen",
         });
       } catch (error) {
         res.status(400).json({ success: false, data: error });
